Replace float matrices with int matrices in MatrixMultiplyTP

+++ Tests/Brahma.FSharp.OpenCL.TypeProvider.Test/OpenCLSources/matmat.cl
@@ -1,7 +1,7 @@
 __kernel void myGEMM1(const int M, const int N, const int K,
-                      const __global float* A,
-                      const __global float* B,
-                      __global float* C) {
+                      const __global int* A,
+                      const __global int* B,
+                      __global int* C) {
     // Thread identifiers
     const int globalRow = get_global_id(0); // Row ID of C (0..M)
     const int globalCol = get_global_id(1); // Col ID of C (0..N)
